fixed learning speed

+++ Lab_4k_1sem/MSSHI/lab8_Perceptron/Perceptrone_UI/Form1.cs
@@ -174,7 +174,7 @@ namespace Perceptrone_UI
                 return;
             }
 
-            int LearnSpeed = Convert.ToInt32(numericUpDown_Learning_speed.Value);
+            double LearnSpeed = Convert.ToDouble(numericUpDown_Learning_speed.Value);
             int CountOfEpochs = Convert.ToInt32(numericUpDown_maxCountOfEpochs.Value);
             myNetwork.Learning_speed = LearnSpeed;
             myNetwork.countOfEpochs = CountOfEpochs;
